move stats notice actions out of jitm
