+++ client/state/jitm/actions.js
@@ -6,7 +6,6 @@ import {
 	JITM_FETCH,
 	JITM_SET,
 	JITM_OPEN_HELP_CENTER,
-	UPDATE_STATS_NOTICE_STATUS_DIRECT,
 } from 'calypso/state/action-types';
 import 'calypso/state/data-layer/wpcom/sites/jitm';
 import 'calypso/state/jitm/init';
@@ -24,20 +23,6 @@ export const dismissJITM = ( siteId, id, featureClass ) => ( {
 	siteId,
 	id,
 	featureClass,
-} );
-
-/**
- * Dismisses a jitm directly
- *
- * @param {string} id The id of the jitm to dismiss
- * @param {string} featureClass The feature class of the jitm to dismiss
- * @returns {Object} The dismiss action
- */
-export const updateStatsNoticeStatusDirect = ( id, featureClass, status = 'dismissed' ) => ( {
-	type: UPDATE_STATS_NOTICE_STATUS_DIRECT,
-	id,
-	featureClass,
-	status,
 } );
 
 /**
